[refactor(folder)] organism -> blocks 변경
