+++ src/components/blocks/BulletinBoard/index.tsx
