+++ src/components/blocks/BulletinBoard/style.scss
