+++ tsconfig.paths.json
@@ -5,7 +5,7 @@
             "~/*": ["./src/*"],
             "@atoms/*": ["./src/components/atoms/*"],
             "@molecules/*": ["./src/components/molecules/*"],
-            "@organisms/*": ["./src/components/organisms/*"],
+            "@blocks/*": ["./src/components/blocks/*"],
             "@lib/*": ["./src/lib/*"]
         }
     }
